// apps/pc/src/generated/models/Pl_contactsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pl_contactspl_statuscode = {
  100000000: '在籍',
  100000001: '休職',
  100000002: '退職'
} as const;
export type Pl_contactspl_statuscode = keyof typeof Pl_contactspl_statuscode;
export const Pl_contactsstatecode = {
  0: 'アクティブ',
  1: '非アクティブ'
} as const;
export type Pl_contactsstatecode = keyof typeof Pl_contactsstatecode;
export const Pl_contactsstatuscode = {
  1: 'アクティブ',
  2: '非アクティブ'
} as const;
export type Pl_contactsstatuscode = keyof typeof Pl_contactsstatuscode;

export interface Pl_contactsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  pl_contactid: string;
  pl_departmentrole?: string;
  pl_email?: string;
  pl_name?: string;
  "pl_PartnerLookup@odata.bind"?: string;
  pl_phone?: string;
  pl_registeredat?: string;
  pl_registrationkey?: string;
  pl_statuscode?: Pl_contactspl_statuscode;
  statecode: Pl_contactsstatecode;
  statuscode?: Pl_contactsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pl_contacts extends Pl_contactsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pl_partnerlookupname?: string;
  pl_statuscodename?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pl_partnerlookup?: object;
  _pl_partnerlookup_value?: string;
}